requirements.txt
streamlit
ultralytics==8.0.196
opencv-python-headless==4.7.0.72
Pillow
numpy